requirements.txt
# SDV Runtime Compatible Requirements
# Includes Velocitas Python SDK for full SDV ecosystem integration

# Core Velocitas SDK
velocitas-sdk==0.9.0

# Additional dependencies for SDV compatibility
python-socketio==5.8.0
requests==2.31.0

# Async runtime and utilities
asyncio-mqtt==0.16.1
aiohttp==3.9.0

# Data validation and processing
pydantic==2.5.0
typing-extensions==4.8.0

# OpenTelemetry for structured logging
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0